refactor: Move ServerConfig to the config folder
