+++ src/lib/gui/config/ServerConfig.h
@@ -7,8 +7,8 @@
 
 #pragma once
 
-#include "Hotkey.h"
 #include "base/NetworkProtocol.h"
+#include "gui/Hotkey.h"
 #include "gui/config/IServerConfig.h"
 #include "gui/config/ScreenConfig.h"
 #include "gui/config/ScreenList.h"
